Add hook point for deleting a process definition
Throw UnsupportedOperationException if a single process definition is deleted through the entity manager.
This never worked, since there is no deleteProcessDefinition statement in the ProcessDefinition.xml MyBatis mapping.
